% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cor_mat}
\alias{cor_mat}
\title{Lower triangle correlation matrix for a hypothetical trial.}
\format{
A data frame with 25 rows of 26 variables
\describe{
\item{short_name}{Short name for outcomes}
\item{label}{Label to be used for each variable}
\item{mort}{Mortality}
\item{ivh_any}{Any IVH}
\item{sev_ivh}{IVH Grade 3-4}
\item{sepsis}{Proven Sepsis}
\item{cld}{Chronic lung disease}
\item{steroids}{Receipt of Steroids for CLD}
\item{pulm_hem}{Chronic Lung Disease}
\item{nec}{Necrotizing Enterocolitis Stage 2 or greater}
\item{sip}{Spontaneous intestinal perforation}
\item{gibleed7days}{GI bleed within 7 days of first pharmacotherapy}
\item{rop3}{Retinopathy stage 3 or greater}
\item{oliguria}{Oliguria}
\item{prot_dev}{Protocol deviation}
\item{rescue}{Rescue treatment}
\item{procpda}{Procedural PDA closure}
\item{any_rx}{Receipt of any Rx for PDA}
\item{dur_mv}{Duration of mechanical ventilation}
\item{dur_rs}{Duration of any respiratory support}
\item{full_feed}{Time to full enteral feed}
\item{reg_bw}{Time to regain birthweight}
\item{dur_hosp}{Duration of hospitalization}
\item{neur_imp_mot}{Mod/Sev Motor Impairment at 18-24 months}
\item{neur_imp_cog}{Mod/Sev Cognitive Impairment at 18-24 months}
\item{neur_imp_lan}{Mod/Sev Language Impairment at 18-24 months}
}
}
\usage{
cor_mat
}
\description{
A dataset containing the lower triangle of a correlation matrix for a
hypothetical trial. Numbers are based on general guesses and nothing
concrete. Meant to give something semi realistic for the purpose of
explanation.
}
\keyword{datasets}
