% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_mu}
\alias{find_mu}
\title{Find value of mu to give desired probability}
\usage{
find_mu(p)
}
\arguments{
\item{p}{Desired probability}
}
\value{
the value of mu to use and a summary of how close it was to the one desired
}
\description{
Find the value of mu from a normal distribution with sd of 1 that would
give the desired probability when event == 1 when x > 0
}
