% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{sim_mv}
\alias{sim_mv}
\title{sim_mv}
\usage{
sim_mv(cor_mat, mus, sds, type)
}
\arguments{
\item{cor_mat}{A correlation matrix. With 1s on the diagonal and correlations
on the off diagonal}

\item{mus}{a vector of means}

\item{type}{a vector of types "c" or "b" for continuous or binary}

\item{ses}{a vector of standard errors}
}
\description{
Take a correlation matrix for mixed binomial and continuous data and simulate
data from a multivariate probit model. A nice introduction to these can be
found here. \insertCite{mcparland2016model}{neoDecision}. The reason I use
this approach is because later I will be simulating data from a model that
assumes this and then trying to recover parameters.
}
\references{
{
\insertAllCited{}
}
}
